Chore: Moved Icons to src/icons
