+++ src/icons/FileIcon.tsx
@@ -1,0 +1,25 @@
+import { createSvgIcon } from '@mui/material'
+import React from 'react'
+
+const FileIcon = createSvgIcon(
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M9.33464 1.33594H4.0013C3.64768 1.33594 3.30854 1.47641 3.05849 1.72646C2.80844 1.97651 2.66797 2.31565 2.66797 2.66927V13.3359C2.66797 13.6896 2.80844 14.0287 3.05849 14.2787C3.30854 14.5288 3.64768 14.6693 4.0013 14.6693H12.0013C12.3549 14.6693 12.6941 14.5288 12.9441 14.2787C13.1942 14.0287 13.3346 13.6896 13.3346 13.3359V5.33594L9.33464 1.33594Z" />
+    <path d="M9.33203 1.33594V5.33594H13.332" />
+    <path d="M10.6654 8.66406H5.33203" />
+    <path d="M10.6654 11.3359H5.33203" />
+    <path d="M6.66536 6H5.9987H5.33203" />
+  </svg>,
+  'FileIcon',
+)
+
+export default FileIcon
