+++ src/icons/ImageIcon.tsx
@@ -1,0 +1,22 @@
+import { createSvgIcon } from '@mui/material'
+
+const ImageIcon = createSvgIcon(
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M12.6667 2H3.33333C2.59695 2 2 2.59695 2 3.33333V12.6667C2 13.403 2.59695 14 3.33333 14H12.6667C13.403 14 14 13.403 14 12.6667V3.33333C14 2.59695 13.403 2 12.6667 2Z" />
+    <path d="M5.66797 6.66406C6.22025 6.66406 6.66797 6.21635 6.66797 5.66406C6.66797 5.11178 6.22025 4.66406 5.66797 4.66406C5.11568 4.66406 4.66797 5.11178 4.66797 5.66406C4.66797 6.21635 5.11568 6.66406 5.66797 6.66406Z" />
+    <path d="M13.9987 9.9974L10.6654 6.66406L3.33203 13.9974" />
+  </svg>,
+  'ImageIcon',
+)
+
+export default ImageIcon
