+++ src/icons/ListIcon.tsx
@@ -1,0 +1,25 @@
+import { createSvgIcon } from '@mui/material'
+
+const ListIcon = createSvgIcon(
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M5.33203 4H13.9987" />
+    <path d="M5.33203 8H13.9987" />
+    <path d="M5.33203 12H13.9987" />
+    <path d="M2 4H2.00667" />
+    <path d="M2 8H2.00667" />
+    <path d="M2 12H2.00667" />
+  </svg>,
+  'ListIcon',
+)
+
+export default ListIcon
