+++ src/icons/TagIcon.tsx
@@ -1,0 +1,22 @@
+import React from 'react'
+import { createSvgIcon } from '@mui/material'
+
+const TagIcon = createSvgIcon(
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M13.7254 8.9426L8.94536 13.7226C8.82153 13.8466 8.67448 13.9449 8.51262 14.012C8.35075 14.0791 8.17725 14.1137 8.00203 14.1137C7.82681 14.1137 7.65331 14.0791 7.49144 14.012C7.32958 13.9449 7.18253 13.8466 7.0587 13.7226L1.33203 8.0026V1.33594H7.9987L13.7254 7.0626C13.9737 7.31242 14.1131 7.65036 14.1131 8.0026C14.1131 8.35485 13.9737 8.69279 13.7254 8.9426Z" />
+    <path d="M4.66797 4.66406H4.67464" />
+  </svg>,
+  'TagIcon',
+)
+
+export default TagIcon
