+++ src/icons/UserCheckIcon.tsx
@@ -1,0 +1,30 @@
+import { createSvgIcon } from '@mui/material'
+import React from 'react'
+
+const UserCheckIcon = createSvgIcon(
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <g clipPath="url(#clip0_3585_14241)">
+      <path d="M10.668 14V12.6667C10.668 11.9594 10.387 11.2811 9.88692 10.781C9.38682 10.281 8.70855 10 8.0013 10H3.33464C2.62739 10 1.94911 10.281 1.44902 10.781C0.94892 11.2811 0.667969 11.9594 0.667969 12.6667V14" />
+      <path d="M5.66667 7.33333C7.13943 7.33333 8.33333 6.13943 8.33333 4.66667C8.33333 3.19391 7.13943 2 5.66667 2C4.19391 2 3 3.19391 3 4.66667C3 6.13943 4.19391 7.33333 5.66667 7.33333Z" />
+      <path d="M11.332 7.33333L12.6654 8.66667L15.332 6" />
+    </g>
+    <defs>
+      <clipPath id="clip0_3585_14241">
+        <rect width="16" height="16" fill="white" />
+      </clipPath>
+    </defs>
+  </svg>,
+  'UserCheckIcon',
+)
+
+export default UserCheckIcon
